Move convention details to shared settings
