+++ client/src/typings/config.typings.ts
@@ -2,9 +2,6 @@ export interface Config {
   SIGNUP_END_TIME: number;
   SIGNUP_OPEN_TIME: number;
   MESSAGE_DELAY: number;
-  CONVENTION_NAME: ConventionName;
-  CONVENTION_YEAR: string;
-  CONVENTION_START_TIME: string;
   DAY_START_TIME: number;
   apiServerUrl: string;
   useTestTime: boolean;
@@ -18,5 +15,3 @@ export interface Config {
   tagFilteringEnabled: boolean;
   simpleDetails: boolean;
 }
-
-type ConventionName = "Ropecon" | "Tracon Hitpoint";
